Revert "chore: update datatype from int to Long in UserProvider"
This reverts commit af38e057e2fc976ac4020dd9a82f27e41bc484fc.

+++ src/main/java/com/yui/providers/UserProvider.java
@@ -28,7 +28,7 @@ public class UserProvider {
     }
 
     @PostMapping("/user")
-    public Long createUser(@RequestBody User user) {
+    public int createUser(@RequestBody User user) {
         return repo.save(user).getId();
     }
